+++ src/browser.detect.js
@@ -1,9 +1,10 @@
 export function _detectBrowser() {
-    var is_firefox = typeof InstallTrigger !== 'undefined';
-    var is_safari = !!window.safari && !!window.ApplePayError;
-    var is_chromium = !!window.chrome && !window.opr;
-    var is_brave = !!navigator.brave && is_chromium;
-    var is_opera = !!window.opr && !!window.chrome;
+    const is_firefox = typeof InstallTrigger !== 'undefined';
+    const is_safari = !!window.safari && !!window.ApplePayError;
+    const is_chromium = !!window.chrome && !window.opr;
+    const is_brave = !!navigator.brave && is_chromium;
+    const is_opera = !!window.opr && !!window.chrome;
+    const is_ddg = (!!window.EmailInterface && !!window.LoginDetection) && !is_chromium;
     /**
      * Currently using 'XRHand' to detect Edge-Chromium,
      * somehow MS Edge has 'XRHand' even tho it is not supported
@@ -13,10 +14,10 @@ export function _detectBrowser() {
      * 
      * TODO: find an better way to check for Edge (Chromium)
      */
-    var is_edgeChrome = !!window.XRHand && (!!window.XRJointPose || !!window.XRJointSpace) && is_chromium;
-    var is_ie = !!window.document.documentMode;
+    const is_edgeChrome = !!window.XRHand && (!!window.XRJointPose || !!window.XRJointSpace) && is_chromium;
+    const is_ie = !!window.document.documentMode;
 
-    var loop_list = [
+    const loop_list = [
         { state: is_brave, id: '7d952f3f' },
         { state: is_opera, id: 'c5bb67fa' },
         { state: is_edgeChrome, id: '050d2b13' },
@@ -24,6 +25,7 @@ export function _detectBrowser() {
         { state: is_firefox, id: '6cc0cdd0' },
         { state: is_safari, id: 'cd2e422a' },
         { state: is_ie, id: 'cb14225b' },
+        { state: is_ddg, id: 'eb2cb20f' }
     ];
 
     var ret = undefined;